# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adil Smart Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductSpecs = { [key: string]: string };

export interface Product {
  id: number;
  name: string;
  slug: string;
  brand: string;
  category: string;
  images: string[];
  description: string;
  specs?: ProductSpecs;
  featured: boolean;
  variants: string[];
  createdAt: string;
}

export interface ProductList {
  products: Product[];
  total: number;
}

export type ProductInputSpecs = { [key: string]: string };

export interface ProductInput {
  name: string;
  slug: string;
  brand: string;
  category: string;
  images: string[];
  description: string;
  specs?: ProductInputSpecs;
  featured?: boolean;
  variants: string[];
}

export type ProductUpdateSpecs = { [key: string]: string };

export interface ProductUpdate {
  name?: string;
  brand?: string;
  category?: string;
  images?: string[];
  description?: string;
  specs?: ProductUpdateSpecs;
  featured?: boolean;
  variants?: string[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon?: string;
  count: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface BrandCount {
  brand: string;
  count: number;
}

export interface ProductStats {
  total: number;
  featured: number;
  byCategory: CategoryCount[];
  byBrand: BrandCount[];
  recentlyAdded: Product[];
}

export type ListProductsParams = {
  category?: string;
  brand?: string;
  search?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
};