# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Adil Smart Store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/featured:
    get:
      operationId: listFeaturedProducts
      tags: [products]
      summary: List featured products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/stats:
    get:
      operationId: getProductStats
      tags: [products]
      summary: Get product statistics for admin dashboard
      responses:
        "200":
          description: Product statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductStats"

  /products/{slug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories with product counts
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      required:
        - id
        - name
        - slug
        - brand
        - category
        - images
        - description
        - featured
        - variants
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        brand:
          type: string
        category:
          type: string
        images:
          type: array
          items:
            type: string
        description:
          type: string
        specs:
          type: object
          additionalProperties:
            type: string
        featured:
          type: boolean
        variants:
          type: array
          items:
            type: string
        createdAt:
          type: string

    ProductList:
      type: object
      required:
        - products
        - total
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer

    ProductInput:
      type: object
      required:
        - name
        - slug
        - brand
        - category
        - images
        - description
        - variants
      properties:
        name:
          type: string
        slug:
          type: string
        brand:
          type: string
        category:
          type: string
        images:
          type: array
          items:
            type: string
        description:
          type: string
        specs:
          type: object
          additionalProperties:
            type: string
        featured:
          type: boolean
        variants:
          type: array
          items:
            type: string

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        category:
          type: string
        images:
          type: array
          items:
            type: string
        description:
          type: string
        specs:
          type: object
          additionalProperties:
            type: string
        featured:
          type: boolean
        variants:
          type: array
          items:
            type: string

    Category:
      type: object
      required:
        - id
        - name
        - slug
        - count
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        count:
          type: integer

    ProductStats:
      type: object
      required:
        - total
        - featured
        - byCategory
        - byBrand
        - recentlyAdded
      properties:
        total:
          type: integer
        featured:
          type: integer
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        byBrand:
          type: array
          items:
            $ref: "#/components/schemas/BrandCount"
        recentlyAdded:
          type: array
          items:
            $ref: "#/components/schemas/Product"

    CategoryCount:
      type: object
      required:
        - category
        - count
      properties:
        category:
          type: string
        count:
          type: integer

    BrandCount:
      type: object
      required:
        - brand
        - count
      properties:
        brand:
          type: string
        count:
          type: integer