# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adil Smart Store API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductSpecs } from "./productSpecs";

export interface Product {
  id: number;
  name: string;
  slug: string;
  brand: string;
  category: string;
  images: string[];
  description: string;
  specs?: ProductSpecs;
  featured: boolean;
  variants: string[];
  createdAt: string;
}