# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adil Smart Store API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputSpecs } from "./productInputSpecs";

export interface ProductInput {
  name: string;
  slug: string;
  brand: string;
  category: string;
  images: string[];
  description: string;
  specs?: ProductInputSpecs;
  featured?: boolean;
  variants: string[];
}