# lib/api-zod/src/generated/types/productList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adil Smart Store API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface ProductList {
  products: Product[];
  total: number;
}