# lib/api-zod/src/generated/types/productStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adil Smart Store API
 * OpenAPI spec version: 0.1.0
 */
import type { BrandCount } from "./brandCount";
import type { CategoryCount } from "./categoryCount";
import type { Product } from "./product";

export interface ProductStats {
  total: number;
  featured: number;
  byCategory: CategoryCount[];
  byBrand: BrandCount[];
  recentlyAdded: Product[];
}