# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adil Smart Store API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductUpdateSpecs } from "./productUpdateSpecs";

export interface ProductUpdate {
  name?: string;
  brand?: string;
  category?: string;
  images?: string[];
  description?: string;
  specs?: ProductUpdateSpecs;
  featured?: boolean;
  variants?: string[];
}